requirements.txt
wheel>=0.42.0
setuptools>=69.0.0
fastapi
uvicorn
python-dotenv
pandas
numpy<2.0.0
requests
python-binance
websockets
ta
google-generativeai
yfinance
feedparser
python-jose[cryptography]
passlib[bcrypt]
python-multipart
textblob
scikit-learn
newsapi-python>=0.2.7
pandas_ta>=0.3.14b
huggingface-hub>=0.21.4
yfinance>=0.1.63
ta-lib==0.4.28
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cpu
google-genai>=0.1.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.3
nltk>=3.8.1
redis>=4.5.1
aioredis>=2.0.1
asyncio>=3.4.3
pytz>=2023.3
scikit-learn>=1.4.0
tensorflow-cpu>=2.15.0  # Using CPU version to avoid GPU dependencies
transformers>=4.37.0
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
python-binance==1.0.19
websockets==12.0
starlette==0.36.3
pydantic==2.6.1
typing_extensions==4.9.0
idna==3.6
sniffio==1.3.0
anyio==4.2.0
click==8.1.7
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
protobuf==4.25.2
regex==2023.12.25
tqdm==4.66.1
decorator==5.1.1
dateparser==1.2.0
tzlocal==5.2
python-dateutil==2.8.2
tzdata==2024.1
backoff==2.2.1
aiohttp==3.9.3
aiosignal==1.3.1
attr==0.3.2
frozenlist==1.4.1
multidict==6.0.4
yarl==1.9.4
feedparser==6.0.11
yfinance>=0.1.63
